requirements.txt
streamlit
tensorflow
scikit-learn
pillow  
#streamlit run app.py